Renaming files to Link.js
